Add comments
- Renamed CustomPass to ExampleComputePass, which better matches what it is
- Added explanatory comments in RenderPass and ExampleComputePass
- Added initial values for RenderPass::m_OutputHandle and ExampleComputePass::m_ExampleComputePipeline
