+++ Example/WorkingDir/Shaders/ExampleCompute.gsh
