+++ Include/Rendering/Frontend/Renderer/RenderPasses/RenderPass.h
@@ -43,6 +43,7 @@ public:
 // Shortened to make inheriting from this type easier
 using BaseConfigData = RenderPassInterface::ConfigDataInterface;
 
+// Inherit from this class using : public RenderPass<DerivedClassName> (or : public Gecko::RenderPass<DerivedClassName>)
 template <typename T>
 class RenderPass : public RenderPassInterface
 {
@@ -50,8 +51,8 @@ public:
 	RenderPass() = default;
 	virtual ~RenderPass() {}
 
-	// This function will cause a runtime error if the object it is called on is not a valid derived class of RenderPass,
-	// if no overridden ConfigData struct exists on the derived class, or if no SubInit function exists on the derived class
+	/* This function will cause a runtime error if the object it is called on is not a valid derived class of RenderPass,
+	   if no overridden ConfigData struct exists on the derived class, or if no SubInit function exists on the derived class */
 	virtual const void Init(const Platform::AppInfo& appInfo, ResourceManager* resourceManager,
 		const BaseConfigData& dependencies) override final
 	{
@@ -62,6 +63,9 @@ public:
 			ASSERT_MSG(false, "Invalid render pass initialisation!");
 	}
 
+	/* Render to output target (pointed to by m_OutputHandle). Any dependencies for rendering should be defined in derived 
+	classes; these are usually set up by defining a ConfigData struct that inherits from BaseConfigData, and passing a 
+	valid ConfigData into the Init() function */
 	virtual const void Render(const SceneRenderInfo& sceneRenderInfo, ResourceManager* resourceManager,
 		const Renderer* renderer, Ref<CommandList> commandList) override = 0;
 
@@ -73,7 +77,7 @@ public:
 protected:
 	// Make sure to write a SubInit function!
 
-	RenderTargetHandle m_OutputHandle;
+	RenderTargetHandle m_OutputHandle{ static_cast<u32>(-1) };
 
 private:
